moves electronic_collections tm generator
