+++ umich_catalog_indexing/lib/jobs.rb
@@ -10,8 +10,8 @@ require "services"
 
 require "jobs/catalog_indexing_metrics"
 require "jobs/utilities"
+require "jobs/translation_map_generator"
 require "jobs/lib_loc_info"
-require "jobs/electronic_collections"
 require "jobs/delete_id_getter"
 require "jobs/delete_alma_ids"
 require "jobs/index_alma_xml"
