+++ umich_catalog_indexing/lib/jobs/translation_map_generator.rb
@@ -1,0 +1,5 @@
+require_relative "translation_map_generator/electronic_collections"
+module Jobs
+  module TranslationMapGenerator
+  end
+end
