+++ umich_catalog_indexing/lib/jobs/utilities/translation_map_fetcher.rb
@@ -6,7 +6,7 @@ module Jobs
         high_level_browse_klass: HighLevelBrowse,
         translation_map_generators: [
           Jobs::LibLocInfo,
-          Jobs::ElectronicCollections
+          Jobs::TranslationMapGenerator::ElectronicCollections
         ],
         translation_map_dir: "/app/lib/translation_maps"
       )
